Tests/LibWeb/Text/expected/SVG/removeChild-on-ancestor-of-use-element.txt

use1 > [object SVGSVGElement] foo
use2 > [object SVGSVGElement] foo
no use1
use2 > [object SVGSVGElement] foo
